Added route for canceling a reservation.
