+++ express_cut/express_service/urls.py
@@ -1,10 +1,11 @@
 from django.urls import path
-from . import views, service_view, role_view
+from . import views, view_service, view_role, view_reservation
 from django.conf.urls import url
 from rest_framework.urlpatterns import format_suffix_patterns
 from rest_framework import permissions
 from drf_yasg.views import get_schema_view
 from drf_yasg import openapi
+
 
 schema_view = get_schema_view(
    openapi.Info(
@@ -24,18 +25,20 @@ urlpatterns = [
     path('schedule', views.schedule_views),
     path('schedule/<int:pk>', views.schedule_views_put),
 
-    path('service', service_view.all_services),
-    path('service/<int:pk>', service_view.service_views),
+    path('service', view_service.all_services),
+    path('service/<int:pk>', view_service.service_views),
 
-    path('reservation', views.reservation_general),
-    path('reservation/<int:pk>', views.reservation_views),
-    path('stylist/<int:stylist_id>/reservation', views.reservations_by_stylist),
+    path('reservation', view_reservation.reservation_general),
+    path('reservation/<int:pk>', view_reservation.reservation_views),
+    path('reservation/<int:pk>/cancel', view_reservation.cancel_reservation),
+
+    path('stylist/<int:stylist_id>/reservation', view_reservation.reservations_by_stylist),
 
   
-    path('stylist', role_view.all_stylists),
-    path('stylist/available', role_view.all_available_stylists),
-    path('customer', role_view.all_customers),
-    path('manager', role_view.all_managers),
+    path('stylist', view_role.all_stylists),
+    path('stylist/available', view_role.all_available_stylists),
+    path('customer', view_role.all_customers),
+    path('manager', view_role.all_managers),
 ]
 
 # Stylist_urlpatterns = [
